--- feature/entity-create-data/build.gradle
@@ -1,11 +1,0 @@
-apply from: "$rootDir/feature_default.gradle"
-android {
-    namespace 'com.krossovochkin.fiberyunofficial.entitycreatedata'
-}
-
-dependencies {
-    api libs.dagger
-    api project(":core:data:api")
-    api project(":core:domain")
-    api project(":feature:entity-create-domain")
-}
