+++ settings.gradle
@@ -31,7 +31,6 @@ include ':feature:entity-list'
 include ':feature:entity-details'
 include ':feature:login'
 include ':feature:entity-create'
-include ':feature:entity-create-data'
 include ':feature:entity-create-domain'
 include ':feature:picker-entity'
 include ':feature:picker-single-select'
